+++ Factory/Factories/GenericFactory.cs
@@ -1,6 +1,6 @@
 using System;
 
-using Factory.Models;
+using Factory.Products;
 
 namespace Factory.Factories
 {
