+++ Factory/Products/HondaTruck.cs
@@ -1,4 +1,4 @@
-namespace Factory.Models
+namespace Factory.Products
 {
     /// <summary>
     /// The truck manufactured by Honda.
