+++ Factory/Products/MazdaCar.cs
@@ -1,9 +1,9 @@
-namespace Factory.Models
+namespace Factory.Products
 {
     /// <summary>
-    /// The truck manufactured by Mazda.
+    /// The car manufactured by Mazda.
     /// </summary>
-    public class MazdaTruck : IVehicle
+    public class MazdaCar : IVehicle
     {
         /// <summary>
         /// The name of the vehicle.
@@ -11,11 +11,11 @@ namespace Factory.Models
         public string Name { get; private set; }
 
         /// <summary>
-        /// Initializes an instance of the <see cref="MazdaTruck" /> class.
+        /// Initializes an instance of the <see cref="MazdaCar" /> class.
         /// </summary>
-        public MazdaTruck()
+        public MazdaCar()
         {
-            Name = "Mazda Truck";
+            Name = "Mazda Car";
         }
     }
 }
